Broke up project.py into diffrent files
