+++ audino/backend/routes/projects.py
@@ -352,307 +352,6 @@ def give_users_examples(user_id):
     )
 
 
-@api.route("/projects/example", methods=["PATCH"])
-def give_users_examples_json():
-    user_id = request.json.get("users")
-    user = User.query.filter_by(username=user_id).first()
-    app.logger.info(f"{user_id}")
-    return give_users_examples(user.id)
-
-
-@api.route("/projects/<int:project_id>/labels", methods=["POST"])
-@jwt_required
-def add_label_to_project(project_id):
-    msg, status, request_user = check_admin_permissions(get_jwt_identity())
-    if (msg is not None):
-        return msg, status
-
-    label_name = request.json.get("name", None)
-    label_type_id = request.json.get("type", None)
-
-    if not label_name:
-        return (
-            jsonify(
-                message="Please provide a label name!",
-                type="LABEL_NAME_MISSING"),
-            400,
-        )
-
-    if not label_type_id:
-        return (
-            jsonify(
-                message="Please provide a label type!",
-                type="LABEL_TYPE_MISSING"),
-            400,
-        )
-
-    label_type_id = int(label_type_id)
-
-    all_label_types = [label_type.id for label_type in LabelType.query.all()]
-
-    if label_type_id not in all_label_types:
-        return (
-            jsonify(
-                message="Please assign correct label type!",
-                type="LABEL_TYPE_INCORRECT"
-            ),
-            400,
-        )
-
-    try:
-        if (len(LabelType.query.all()) == 0):
-            select = LabelType(id=1, type='Select')
-
-            mutliselect = LabelType(id=2, type='Multi-select')
-            db.session.add_all([select, mutliselect])
-            db.session.commit()
-    except Exception as e:
-        app.logger.error(e)
-
-    try:
-        project = Project.query.get(project_id)
-        label_id = 1
-        if len(Label.query.all()) > 0:
-            last_label = Label.query.all()[len(Label.query.all())-1]
-            label_id = last_label.id+1
-        label = Label(name=label_name, type_id=label_type_id, id=label_id)
-        project.labels.append(label)
-        db.session.add(project)
-        db.session.commit()
-        db.session.refresh(label)
-    except Exception as e:
-        if type(e) == sa.exc.IntegrityError:
-            app.logger.error(e)
-            return (
-                jsonify(
-                    message=f"Label: {label_name} already exists!",
-                    type="DUPLICATE_LABEL",
-                ),
-                409,
-            )
-        app.logger.error(f"Error adding label to project: {project_id}")
-        app.logger.error(e)
-        return (
-            jsonify(
-                message=f"Error adding label to project: {project_id}",
-                type="LABEL_CREATION_FAILED",
-            ),
-            500,
-        )
-
-    return (
-        jsonify(
-            project_id=project.id,
-            label_id=label.id,
-            message=f"Label assigned to project: {project.name}",
-            type="LABEL_ASSIGNED_TO_PROJECT",
-        ),
-        201,
-    )
-
-
-@api.route("/projects/<int:project_id>/labels/<int:label_id>", methods=["GET"])
-@jwt_required
-def get_label_for_project(project_id, label_id):
-    identity = get_jwt_identity()
-    request_user = User.query.filter_by(username=identity["username"]).first()
-    is_admin = True if request_user.role.role == "admin" else False
-    p_id = project_id
-    if is_admin is False:
-        return jsonify(message="Unauthorized access!"), 401
-    err_msg = f"No label exists w/ Label ID: {label_id} Project ID: {p_id}"
-    try:
-        label = Label.query.filter_by(
-                                      id=label_id,
-                                      project_id=project_id
-        ).first()
-    except Exception as e:
-        app.logger.error(err_msg)
-        app.logger.error(e)
-        return (
-            jsonify(
-                message=err_msg
-            ),
-            404,
-        )
-
-    return (
-        jsonify(
-            project_id=project_id,
-            label_id=label.id,
-            label_name=label.name,
-            label_type_id=label.label_type.id,
-            label_type=label.label_type.type,
-            created_on=label.created_at.strftime("%B %d, %Y"),
-        ),
-        200,
-    )
-
-
-@api.route("/projects/<int:project_id>/labels/<int:label_id>",
-           methods=["PATCH"])
-@jwt_required
-def update_label_for_project(project_id, label_id):
-    msg, status, request_user = check_admin_permissions(get_jwt_identity())
-    if (msg is not None):
-        return msg, status
-
-    label_type_id = request.json.get("type", None)
-    label_name = request.json.get("name", None)
-
-    if not label_type_id:
-        return (
-            jsonify(
-                message="Please provide valid label type!",
-                type="LABEL_TYPE_MISSING"
-            ),
-            400,
-        )
-
-    label_type_id = int(label_type_id)
-    all_label_types = [label_type.id for label_type in LabelType.query.all()]
-
-    if label_type_id not in all_label_types:
-        return (
-            jsonify(
-                message="Please assign correct label type!",
-                type="LABEL_TYPE_INCORRECT"
-            ),
-            400,
-        )
-
-    p_id = project_id
-    err_msg = f"No label exists w/ Label ID: {label_id} Project ID: {p_id}"
-    try:
-        label = Label.query.filter_by(
-                                      id=label_id,
-                                      project_id=project_id
-        ).first()
-        label.set_label_type(label_type_id)
-        label.set_label_name(label_name)
-        db.session.commit()
-    except Exception as e:
-        # TODO: Check for errors here
-        app.logger.error(err_msg)
-        app.logger.error(e)
-        return (
-            jsonify(err_msg), 404,
-        )
-
-    return (
-        jsonify(
-            project_id=project_id,
-            label_id=label.id,
-            label_name=label.name,
-            label_type_id=label.label_type.id,
-            label_type=label.label_type.type,
-            created_on=label.created_at.strftime("%B %d, %Y"),
-        ),
-        200,
-    )
-
-
-@api.route("/projects/<int:project_id>/labels", methods=["GET"])
-@jwt_required
-def get_labels_for_project(project_id):
-    identity = get_jwt_identity()
-
-    try:
-        request_user = User.query.filter_by(
-                                            username=identity["username"]
-        ).first()
-        project = Project.query.get(project_id)
-
-        if request_user not in project.users:
-            return jsonify(message="Unauthorized access!"), 401
-
-        labels = project.labels
-
-        response = {}
-        for label in labels:
-            values = label.label_values
-            type = label.label_type.type
-
-            values = [{"value_id": value.id, "value": value.value}
-                      for value in values]
-
-            response[label.name] = {
-                "type": type,
-                "label_id": label.id,
-                "values": values,
-            }
-
-    except Exception as e:
-        app.logger.error("Error fetching all labels")
-        app.logger.error(e)
-        return (jsonify(message="Error fetching all labels"), 404)
-
-    return (jsonify(response), 200)
-
-
-@api.route("/projects/<int:project_id>/data/<int:data_id>", methods=["GET"])
-@jwt_required
-def get_segmentations_for_data(project_id, data_id):
-    identity = get_jwt_identity()
-
-    try:
-        request_user = User.query.filter_by(
-                                            username=identity["username"]
-        ).first()
-        project = Project.query.get(project_id)
-
-        if request_user not in project.users:
-            msg = "Unauthorized access! User not in project users"
-            return jsonify(message=msg), 401
-
-        data = Data.query.filter_by(id=data_id, project_id=project_id).first()
-
-        segmentations = []
-        for segment in data.segmentations:
-            resp = {
-                "segmentation_id": segment.id,
-                "start_time": segment.start_time,
-                "end_time": segment.end_time,
-                "max_freq": segment.max_freq,
-                "min_freq": segment.min_freq,
-            }
-
-            values = dict()
-            for value in segment.values:
-                if value.label.name not in values:
-                    values[value.label.name] = {
-                        "label_id": value.label.id,
-                        "values": []
-                        if value.label.label_type.type == "Multi-select"
-                        else None,
-                    }
-
-                if value.label.label_type.type == "Multi-select":
-                    values[value.label.name]["values"].append(value.id)
-                else:
-                    values[value.label.name]["values"] = value.id
-
-            resp["annotations"] = values
-
-            segmentations.append(resp)
-
-        response = {
-            "filename": data.filename,
-            "original_filename": data.original_filename,
-            "is_marked_for_review": data.is_marked_for_review,
-            "segmentations": segmentations,
-            "sampling_rate": data.sampling_rate,
-            "clip_length": data.clip_length,
-        }
-
-    except Exception as e:
-        app.logger.error("Error fetching datapoint with given id")
-        app.logger.error(e)
-        return (jsonify(message="Error fetching datapoint with given id"), 404)
-
-    return (jsonify(response), 200)
-
-
 @api.route("/projects/<int:project_id>/data/<int:data_id>", methods=["PATCH"])
 @jwt_required
 def update_data(project_id, data_id):
@@ -701,161 +400,6 @@ def update_data(project_id, data_id):
             type="DATA_UPDATED",
         ),
         200,
-    )
-
-
-@api.route(
-    "/projects/<int:project_id>/data/<int:data_id>/segmentations",
-    methods=["POST"]
-)
-@api.route(
-    "/projects/<int:project_id>/data/<int:data_id>/segmentations/<int:seg_id>",
-    methods=["PUT"],
-)
-@jwt_required
-def add_segmentations(project_id, data_id, seg_id=None):
-    identity = get_jwt_identity()
-    segmentation_id = seg_id
-    if not request.is_json:
-        return jsonify(message="Missing JSON in request"), 400
-
-    start_time = float(request.json.get("start", None))
-    end_time = float(request.json.get("end", None))
-    max_freq = float(request.json.get("regionTopFrequency", None))
-    min_freq = float(request.json.get("regionBotFrequency", None))
-
-    if start_time is None or end_time is None:
-        return (
-            jsonify(message="Params `start_time` or `end_time` missing"), 400
-        )
-
-    if max_freq is None or min_freq is None:
-        return jsonify(message="Params `max_freq` or `min_freq` missing"), 400
-
-    if type(start_time) is not float or type(end_time) is not float:
-        msg = "Params `start_time` and `end_time` need to be float values"
-        return (
-            jsonify(
-                message=msg
-            ),
-            400,
-        )
-
-    if type(max_freq) is not float or type(min_freq) is not float:
-        return (
-            jsonify(
-                message="Params `max_freq` and `min_freq` need to be floats"
-            ),
-            400,
-        )
-
-    annotations = request.json.get("annotations", dict())
-    # miliseconds to seconds
-    time_spent = request.json.get("time_spent", 0) / 1000
-    start_time = round(start_time, 4)
-    end_time = round(end_time, 4)
-    username = identity["username"]
-    max_freq = round(max_freq, 4)
-    min_freq = round(min_freq, 4)
-    try:
-        request_user = User.query.filter_by(username=identity["username"]
-                                            ).first()
-        project = Project.query.get(project_id)
-
-        if request_user not in project.users:
-            return jsonify(message="Unauthorized access!"), 401
-
-        data = Data.query.filter_by(id=data_id, project_id=project_id).first()
-        # if request_user.username not in  data.assigned_user_id:
-        #    return jsonify(message="Unauthorized access!"), 401
-
-        segmentation = generate_segmentation(
-            data_id=data_id,
-            project_id=project.id,
-            end_time=end_time,
-            start_time=start_time,
-            max_freq=max_freq,
-            min_freq=min_freq,
-            annotations=annotations,
-            time_spent=time_spent,
-            username=username,
-            segmentation_id=segmentation_id
-        )
-        db.session.add(segmentation)
-        db.session.commit()
-        db.session.refresh(segmentation)
-    except Exception as e:
-        app.logger.error(f"Could not create segmentation")
-        app.logger.error(e)
-        return (
-            jsonify(
-                message=f"Could not create segmentation",
-                type="SEGMENTATION_CREATION_FAILED",
-            ),
-            500,
-        )
-
-    if request.method == "POST":
-        message = "Segmentation created"
-        operation_type = "SEGMENTATION_CREATED"
-        status = 201
-    else:
-        message = "Segmentation updated"
-        operation_type = "SEGMENTATION_UPDATED"
-        status = 204
-
-    return (
-        jsonify(segmentation_id=segmentation.id, message=message,
-                type=operation_type),
-        status,
-    )
-
-
-@api.route(
-    "/projects/<int:project_id>/data/<int:data_id>/segmentations/<int:seg_id>",
-    methods=["DELETE"],
-)
-@jwt_required
-def delete_segmentations(project_id, data_id, seg_id):
-    identity = get_jwt_identity()
-    segmentation_id = seg_id
-    try:
-        request_user = User.query.filter_by(username=identity["username"]
-                                            ).first()
-        project = Project.query.get(project_id)
-
-        if request_user not in project.users:
-            return jsonify(message="Unauthorized access!"), 401
-
-        data = Data.query.filter_by(id=data_id, project_id=project_id).first()
-
-        # if request_user.username not in  data.assigned_user_id:
-        #    return jsonify(message="Unauthorized access!"), 401
-
-        segmentation = Segmentation.query.filter_by(
-            data_id=data_id, id=segmentation_id
-        ).first()
-
-        db.session.delete(segmentation)
-        db.session.commit()
-    except Exception as e:
-        app.logger.error(f"Could not delete segmentation")
-        app.logger.error(e)
-        return (
-            jsonify(
-                message=f"Could not delete segmentation",
-                type="SEGMENTATION_DELETION_FAILED",
-            ),
-            500,
-        )
-
-    return (
-        jsonify(
-            segmentation_id=segmentation_id,
-            message="Segmentation deleted",
-            type="SEGMENTATION_DELETED",
-        ),
-        204,
     )
